requirements.txt
flask
jinja2
hugchat
huggingchat
reportlab
python-docx==1.1.0
typing-extensions==4.8.0